Add files via upload
You have N cards placed in front of you on the table. The i th card has the number Ai written on it.

In one move, you can remove any one card from the remaining cards on the table.

Find the minimum number of moves required so that all the cards remaining on the table have the same number written on them.