Add files via upload
Your task is to delete a subsequence of maximum length from the string, such that, after concatenating the remaining parts of the string, it becomes a palindrome of length greater than 1.

If this is possible, print the maximum length of the subsequence that can be deleted. Otherwise, print −1.

Input Format
The first line of input will contain a single integer T, denoting the number of test cases.
Each test case consists of 2 lines of input:
The first line consists the a single integer N - the length of string S.
The second line contains string S, consisting of lowercase english alphabets.