Add files via upload
In the bustling CodeChef office, the entrance is equipped with a high-tech card swipe system. Each employee is assigned a unique ID card that they use to swipe in and out of the building.

The system records every swipe, capturing the first swipe of an ID as in, second as out, third as in, and so on. Given an array A consisting of N IDs denoting N swipes throughout the day, find the maximum number of people in the office at any time.

Note that there is nobody inside the office before the first swipe.

Input Format
The first line of input will contain a single integer T, denoting the number of test cases.
Each test case consists of multiple lines of input.
The first line of each test case contains an integer N — the number of swipes throughout the day.
The next line contains N space-separated integers denoting the ID of the ith swipe.
Output Format
For each test case, output on a new line, the maximum number of people in the office at any time.