Add files via upload
In a recent breakthrough in mathematics, the proof utilized a concept called Height.

Consider a fraction a/b. Its Height is defined as the maximum of its numerator and denominator. So, for example, the Height of 3/19 would be 19, and the Height of 27/4 would be 27.

Given 
a and b, find the Height of a/b .

Input Format
The only line of input contains two integers, a and b.

Output Format
Output a single integer, which is the Height of a/b .

Constraints
1≤a,b≤100
a and b do not have any common factors.
Sample 1:
Input
3 19
Output
19
Explanation:
The maximum of 
{3,19} is 19. Hence the Height of 3/19 is 19.

Sample 2:
Input
27 4
Output
27
Explanation:
The maximum of {27,4} is 27. Hence the Height of 27/4 is 27.