Add files via upload
In the light of G-20 summit, government has decided to keep the average air quality index (AQI) strictly below 100.
On some random day, Chef measures the AQI and found the value to be X.

Find whether the government was able to keep the AQI within limits.

Input Format
The input consists of an integer X — the AQI Chef measured.
Output Format
Output YES, if the government was able to keep the AQI within limits and NO otherwise.

You may print each character of the string in uppercase or lowercase (for example, the strings YES, yEs, yes, and yeS will all be treated as identical).

Constraints
1≤X≤150
Sample 1:
Input
50
Output
YES
Explanation:
The AQI is strictly less than 100. Thus, the government was able to keep the AQI within limits.

Sample 2:
Input
100
Output
NO
Explanation:
The AQI is equal to 100. Thus, the government was not able to keep the AQI within limits.

Sample 3:
Input
99
Output
YES
Explanation:
The AQI is strictly less than 100. Thus, the government was able to keep the AQI within limits.

Sample 4:
Input
145
Output
NO
Explanation:
The AQI is greater than 100. Thus, the government was not able to keep the AQI within limits.